Convert all tests
